requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
black==20.8b1
chardet==4.0.0
click==7.1.2
discord.py==1.6.0
idna==3.1
isort==5.8.0
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.7.0
regex==2021.3.17
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
yarl==1.6.3
